+++ CHANGELOG.md
@@ -5,10 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-10-14
+
+## Add
+
+- Numero de ticket editable y funcional.
+- Boton para incrementar el numero de ticket y limpiar el formulario.
+- Numero de ticket persistente en el localstorage.
 
 ## [0.1.0] - 2023-10-14
 
 ## Add
+
 - Columna importe en el ticket
 - Consumidor Final como cabecera del ticket.
 - Muchas gracias en el footer del ticket.
@@ -18,4 +26,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.1] - 2023-10-01
 
 ## Add
+
 - Electron app with NestJS in the backend, react in the frontend along with mui/joy components framework.